requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
faker==25.2.0
streamlit==1.35.0
plotly==5.22.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
gspread==6.1.2
oauth2client==4.1.3
pytest==8.2.2
pytest-mock==3.14.0
